refactor: decompose MLXBackend
